requirements/dev.txt
-r common.txt

# development only dependencies in aphabetical order
django-debug-toolbar>=1.11
httpie>=1.0.2
pytest-django>=3.5.0
pytest-cov>=2.7
pytest-dotenv
